Unify server service files
