+++ server/src/features/settings/settingsService.ts
@@ -1,7 +1,19 @@
-import { findSettings } from 'server/features/settings/settingsRepository';
+import {
+  findSettings,
+  saveSignupTime,
+  saveToggleAppOpen,
+  saveHidden,
+} from 'server/features/settings/settingsRepository';
 import { logger } from 'server/utils/logger';
 import { ServerError } from 'shared/typings/api/errors';
-import { GetSettingsResponse } from 'shared/typings/api/settings';
+import {
+  GetSettingsResponse,
+  PostHiddenResponse,
+  PostToggleAppOpenResponse,
+} from 'shared/typings/api/settings';
+import { PostSignupTimeResponse } from 'shared/typings/api/signup';
+import { Game } from 'shared/typings/models/game';
+import { removeHiddenGamesFromUsers } from 'server/features/settings/settingsUtils';
 
 export const fetchSettings = async (): Promise<
   GetSettingsResponse | ServerError
@@ -24,4 +36,75 @@ export const fetchSettings = async (): Promise<
       code: 0,
     };
   }
+};
+
+export const toggleAppOpen = async (
+  appOpen: boolean
+): Promise<PostToggleAppOpenResponse | ServerError> => {
+  try {
+    const response = await saveToggleAppOpen(appOpen);
+    return {
+      message: 'Update app open success',
+      status: 'success',
+      appOpen: response.appOpen,
+    };
+  } catch (error) {
+    return {
+      message: 'Update app open failure',
+      status: 'error',
+      code: 0,
+    };
+  }
+};
+
+export const storeSignupTime = async (
+  signupTime: string
+): Promise<PostSignupTimeResponse | ServerError> => {
+  try {
+    const response = await saveSignupTime(signupTime);
+    return {
+      message: 'Signup time set success',
+      status: 'success',
+      signupTime: response.signupTime,
+    };
+  } catch (error) {
+    return {
+      message: 'Signup time set failure',
+      status: 'error',
+      code: 0,
+    };
+  }
+};
+
+export const storeHidden = async (
+  hiddenData: readonly Game[]
+): Promise<PostHiddenResponse | ServerError> => {
+  let settings;
+  try {
+    settings = await saveHidden(hiddenData);
+  } catch (error) {
+    logger.error(`saveHidden error: ${error}`);
+    return {
+      message: 'Update hidden failure',
+      status: 'error',
+      code: 0,
+    };
+  }
+
+  try {
+    await removeHiddenGamesFromUsers(settings.hiddenGames);
+  } catch (error) {
+    logger.error(`removeHiddenGamesFromUsers error: ${error}`);
+    return {
+      message: 'Update hidden failure',
+      status: 'error',
+      code: 0,
+    };
+  }
+
+  return {
+    message: 'Update hidden success',
+    status: 'success',
+    hiddenGames: settings.hiddenGames,
+  };
 };
