+++ server/src/features/settings/settingsUtils.ts
@@ -1,44 +1,8 @@
+import { findUsers, updateUser } from 'server/features/user/userRepository';
 import { logger } from 'server/utils/logger';
 import { Game } from 'shared/typings/models/game';
-import { saveHidden } from 'server/features/settings/settingsRepository';
-import { findUsers, updateUser } from 'server/features/user/userRepository';
-import { ServerError } from 'shared/typings/api/errors';
-import { PostHiddenResponse } from 'shared/typings/api/settings';
 
-export const storeHidden = async (
-  hiddenData: readonly Game[]
-): Promise<PostHiddenResponse | ServerError> => {
-  let settings;
-  try {
-    settings = await saveHidden(hiddenData);
-  } catch (error) {
-    logger.error(`saveHidden error: ${error}`);
-    return {
-      message: 'Update hidden failure',
-      status: 'error',
-      code: 0,
-    };
-  }
-
-  try {
-    await removeHiddenGamesFromUsers(settings.hiddenGames);
-  } catch (error) {
-    logger.error(`removeHiddenGamesFromUsers error: ${error}`);
-    return {
-      message: 'Update hidden failure',
-      status: 'error',
-      code: 0,
-    };
-  }
-
-  return {
-    message: 'Update hidden success',
-    status: 'success',
-    hiddenGames: settings.hiddenGames,
-  };
-};
-
-const removeHiddenGamesFromUsers = async (
+export const removeHiddenGamesFromUsers = async (
   hiddenGames: readonly Game[]
 ): Promise<void> => {
   logger.info(`Remove hidden games from users`);
